// .prisma-client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  USER: 'USER',
  PSYCHOLOGIST: 'PSYCHOLOGIST'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const SessionStatus = {
  SCHEDULED: 'SCHEDULED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus]


export const RiskLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH'
} as const

export type RiskLevel = (typeof RiskLevel)[keyof typeof RiskLevel]